feat(remnawave): compatibility with panel v2.8.0
- HwidDevice: rename UserUuid→UserId (int64), add RequestIp (*string) to match v2.8.0 response
- TrafficLimitBytes, UsedTrafficBytes, LifetimeUsedTrafficBytes: int→int64 across all structs and callers
- Remove redundant int64() casts and overflow guard after type promotion
- Update Version Support table in readme for Remnawave 2.8.*

+++ cmd/app/main.go
@@ -658,10 +658,10 @@ func checkCardlinkTopups(
 			continue
 		}
 
-		targetBytes := int64(rwUser.TrafficLimitBytes) + int64(topup.GBAmount)*int64(config.BytesInGigabyte())
+		targetBytes := rwUser.TrafficLimitBytes + int64(topup.GBAmount)*int64(config.BytesInGigabyte())
 		rwUUID := rwUser.UUID
 
-		if err := remnawaveClient.UpdateUserTrafficLimit(ctx, rwUUID, int(targetBytes), rwUser.TrafficLimitStrategy); err != nil {
+		if err := remnawaveClient.UpdateUserTrafficLimit(ctx, rwUUID, targetBytes, rwUser.TrafficLimitStrategy); err != nil {
 			if err := topupRepository.MarkFailed(ctx, topup.ID); err != nil {
 				slog.Error("cardlink topup: mark failed after rw update", "topup_id", topup.ID, "error", err)
 			}
