+++ internal/payment/payment.go
@@ -98,7 +98,7 @@ func (s PaymentService) ProcessPurchaseById(ctx context.Context, purchaseId int6
 	// Keep totalTrafficLimit as int64 until the API call that requires int.
 	totalTrafficLimitBytes := int64(config.TrafficLimit()) + rolloverBytes
 
-	user, err := s.remnawaveClient.CreateOrUpdateUser(ctx, customer.ID, customer.TelegramID, int(totalTrafficLimitBytes), purchase.Month*config.DaysInMonth(), false)
+	user, err := s.remnawaveClient.CreateOrUpdateUser(ctx, customer.ID, customer.TelegramID, totalTrafficLimitBytes, purchase.Month*config.DaysInMonth(), false)
 	if err != nil {
 		return err
 	}
@@ -200,7 +200,7 @@ func (s PaymentService) ProcessPurchaseById(ctx context.Context, purchaseId int6
 	if err != nil {
 		return err
 	}
-	refereeUser, err := s.remnawaveClient.CreateOrUpdateUser(ctxReferee, refereeCustomer.ID, refereeCustomer.TelegramID, config.TrafficLimit(), config.GetReferralDays(), false)
+	refereeUser, err := s.remnawaveClient.CreateOrUpdateUser(ctxReferee, refereeCustomer.ID, refereeCustomer.TelegramID, int64(config.TrafficLimit()), config.GetReferralDays(), false)
 	if err != nil {
 		return err
 	}
@@ -280,7 +280,7 @@ func (s PaymentService) CancelTributePurchase(ctx context.Context, telegramId in
 	if tributePurchase == nil {
 		return errors.New("tribute purchase not found")
 	}
-	expireAt, err := s.remnawaveClient.DecreaseSubscription(ctx, telegramId, config.TrafficLimit(), -tributePurchase.Month*config.DaysInMonth())
+	expireAt, err := s.remnawaveClient.DecreaseSubscription(ctx, telegramId, int64(config.TrafficLimit()), -tributePurchase.Month*config.DaysInMonth())
 	if err != nil {
 		return err
 	}
@@ -439,7 +439,7 @@ func (s PaymentService) ActivateTrial(ctx context.Context, telegramId int64) (st
 	if customer == nil {
 		return "", fmt.Errorf("customer %d not found", telegramId)
 	}
-	user, err := s.remnawaveClient.CreateOrUpdateUser(ctx, customer.ID, telegramId, config.TrialTrafficLimit(), config.TrialDays(), true)
+	user, err := s.remnawaveClient.CreateOrUpdateUser(ctx, customer.ID, telegramId, int64(config.TrialTrafficLimit()), config.TrialDays(), true)
 	if err != nil {
 		slog.Error("Error creating user", "error", err)
 		return "", err
@@ -535,7 +535,7 @@ func (s PaymentService) calculateRollover(ctx context.Context, telegramID int64)
 		return topupBytes
 	}
 
-	usedBeyondBase := int64(u.UserTraffic.UsedTrafficBytes) - baseBytes
+	usedBeyondBase := u.UserTraffic.UsedTrafficBytes - baseBytes
 	if usedBeyondBase <= 0 {
 		return topupBytes
 	}
